requirements.txt
matplotlib>=2.0.0
ipython>=5.3.0
numpy>=1.6.1
scipy>=0.19.0
scikit-learn>=0.18.1
http://download.pytorch.org/whl/cu80/torch-0.4.1-cp35-cp35m-linux_x86_64.whl
tqdm>=4.19.8
seaborn>=0.8.1
scikit-image>=0.14.0
torchvision>=0.2.1
pillow==5.0.0